requirements.txt
certifi
fastapi
groq
huggingface-hub
ipykernel
langchain
langchain-community
langchain-core
langchain-experimental
langchain-groq
langchain-huggingface
langchain-neo4j
langchain-ollama
langchain-openai
langchain-text-splitters
langchainhub
langgraph
langgraph-checkpoint
langgraph-sdk
langsmith
matplotlib-inline
mypy-extensions
mysql-connector-python
neo4j
nest-asyncio
numpy
ollama
openai
orjson
pandas
pillow
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
pyodbc
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
regex
requests
requests-toolbelt
scikit-learn
scipy
sentence-transformers
setuptools
SQLAlchemy
streamlit
tenacity
tiktoken
tokenizers
toml
torch
tqdm
traitlets
transformers
types-requests
typing-inspect
typing_extensions
urllib3
uvicorn
spacy
rank-bm25